usermanagement/requirements.txt
Django==2.0
django-extensions==2.2.8
djangorestframework==3.11.0
postgis==1.0.4
psycopg2-binary==2.8.4
python-http-client==3.2.6
pytz==2019.3
sendgrid==6.2.0
six==1.14.0
